- CRAI suggestions, Rename folders to reflect component names,  version bump to 0.4.0
