+++ .storybook/manager.js
@@ -1,4 +1,4 @@
-import { addons } from 'storybook/manager-api';
+import { addons } from '@storybook/manager-api';
 import { themes } from 'storybook/theming';
 
 addons.setConfig({
