+++ saiy2k-nostr-components/README.md
@@ -16,6 +16,7 @@ wp-plugin/
 │   └── Kses.php              # Security (custom elements)
 ├── blocks/                   # Gutenberg block definitions
 │   ├── nostr-zap-button/
+│   ├── nostr-like-button/
 │   ├── nostr-follow-button/
 │   ├── nostr-post/
 │   ├── nostr-profile/
