+++ saiy2k-nostr-components/README.txt
@@ -4,7 +4,7 @@ Tags: nostr, social, blocks, gutenberg, shortcodes, zap, like
 Requires at least: 6.0
 Tested up to: 6.8
 Requires PHP: 7.4
-Stable tag: 0.3.2
+Stable tag: 0.4.0
 License: MIT
 License URI: https://opensource.org/licenses/MIT
