+++ saiy2k-nostr-components/blocks/nostr-like-button/block.json
