+++ saiy2k-nostr-components/blocks/nostr-like-button/index.js
