+++ saiy2k-nostr-components/blocks/nostr-zap-button/block.json
