+++ src/index.ts
@@ -5,33 +5,33 @@ import './nostr-profile-badge/nostr-profile-badge.ts';
 import './nostr-post/nostr-post.ts';
 import './nostr-profile/nostr-profile.ts';
 import './nostr-follow-button/nostr-follow-button.ts';
-import './nostr-zap/nostr-zap.ts';
+import './nostr-zap-button/nostr-zap.ts';
 import './nostr-comment/nostr-comment.ts';
 import './nostr-dm/nostr-dm.ts';
 import './nostr-live-chat/nostr-live-chat.ts';
-import './nostr-like/nostr-like.ts';
+import './nostr-like-button/nostr-like.ts';
 
 // Import classes for default export
 import NostrProfileBadge from './nostr-profile-badge/nostr-profile-badge.ts';
 import NostrPost from './nostr-post/nostr-post.ts';
 import NostrProfile from './nostr-profile/nostr-profile.ts';
 import NostrFollowButton from './nostr-follow-button/nostr-follow-button.ts';
-import NostrZap from './nostr-zap/nostr-zap.ts';
+import NostrZap from './nostr-zap-button/nostr-zap.ts';
 import NostrComment from './nostr-comment/nostr-comment.ts';
 import NostrDm from './nostr-dm/nostr-dm.ts';
 import NostrLiveChat from './nostr-live-chat/nostr-live-chat.ts';
-import NostrLike from './nostr-like/nostr-like.ts';
+import NostrLike from './nostr-like-button/nostr-like.ts';
 
 // Export classes for TypeScript types
 export { default as NostrProfileBadge } from './nostr-profile-badge/nostr-profile-badge.ts';
 export { default as NostrPost } from './nostr-post/nostr-post.ts';
 export { default as NostrProfile } from './nostr-profile/nostr-profile.ts';
 export { default as NostrFollowButton } from './nostr-follow-button/nostr-follow-button.ts';
-export { default as NostrZap } from './nostr-zap/nostr-zap.ts';
+export { default as NostrZap } from './nostr-zap-button/nostr-zap.ts';
 export { default as NostrComment } from './nostr-comment/nostr-comment.ts';
 export { default as NostrDm } from './nostr-dm/nostr-dm.ts';
 export { default as NostrLiveChat } from './nostr-live-chat/nostr-live-chat.ts';
-export { default as NostrLike } from './nostr-like/nostr-like.ts';
+export { default as NostrLike } from './nostr-like-button/nostr-like.ts';
 
 // Export init function for explicit initialization (optional, for compatibility)
 export function init() {
