+++ src/nostr-like-button/dialog-help.ts
