+++ src/nostr-like-button/like-utils.ts
