+++ src/nostr-like-button/render.ts
