+++ src/nostr-like-button/spec/implementation.md
