+++ src/nostr-like-button/spec/spec.md
