+++ src/nostr-like-button/spec/testing.md
