+++ src/nostr-zap-button/dialog-help-style.ts
