+++ src/nostr-zap-button/dialog-help.ts
