+++ src/nostr-zap-button/dialog-zap.ts
