+++ src/nostr-zap-button/render.ts
