+++ src/nostr-zap-button/spec/wireframe.png
